+++ liste_arete.h
@@ -1,7 +1,7 @@
 //liste_arete.h
 
 struct maillon_arete{
-  int poids;
+  double poids;
   int idxSommet;
   struct maillon_arete* suivant;
 };
@@ -14,7 +14,7 @@ struct liste_arete{
 };
 
 extern void init_liste_arete(struct liste_arete*);
-extern void ajout_en_queue_arete(int idx, int cost, struct liste_arete* L);
+extern void ajout_en_queue_arete(int idx, double cost, struct liste_arete* L);
 extern void affiche_liste_arete(struct liste_arete*);
 extern struct maillon_arete* minimum(struct liste_arete*);
 extern void clear_liste_arete(struct liste_arete*);
